# openapi.yaml
openapi: 3.0.3
info:
  title: INFO6150 - Web Design and User Experience Engineering Final Project || LEETIFY
  description: 
    Leetify is an online platform that provides a collection of coding challenges to help developers prepare for technical interviews.
  version: "1.0.0"
...
servers:
  - url: http://localhost:8000/
...
paths:
  /users:
    get:
      tags:
        - users
      summary: Get all users
      description: Get all users
      operationId: getallusers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal server error
  /users/{id}:
    get:
      tags:
        - users
      summary: Get a user
      description: Get a user
      operationId: getauser
      parameters:
        - name: id
          in: path
          description: 'Fetch user by ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Get a user
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal server error
    put:
      tags:
        - users
      summary: Update user details
      description: Update user details
      operationId: updateuser
      parameters:
        - name: id
          in: path
          description: 'Update user'
          required: true
          schema:
            type: integer
      requestBody:
        description: Register user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: Update user
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/User'
        '404':
          description: User not found!
        '500':
          description: Internal server error
    delete:
      tags:
        - users
      summary: Delete user details
      description: Delete user details
      operationId: deleteuser
      parameters:
        - name: id
          in: path
          description: 'Delete user'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details deleted!
        '404':
          description: User not found!
        '500':
          description: Internal server error
...
components:
  schemas:
...
    Submission:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 11
        problem_id:
          type: integer
          format: int64
          example: 7
        user_id:
          type: integer
          format: int64
          example: 10
        submitted_time:
          type: string
        language:
          type: string
          example: Java
        submission:
          type: object
        status:
          type: string
          example: pending
          enum:
            - pending
            - accepted
            - rejected
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: username
        email:
          type: string
          example: example@gmail.com
        password:
          type: string
          example: password
        phone:
          type: string
          example: 8775429987
        role:
          type: string
          example: user
          enum:
            - user
            - premium_user
            - admin
        score:
          type: integer
          example: 89
        problemsSolved:
          type: integer
          example: 65
        submissions:
          type: array
          items:
            $ref: '#/components/schemas/Submission'